+++ simrec/datasets/dataset.py
@@ -16,20 +16,13 @@
 import os
 import cv2
 import json, re, en_vectors_web_lg, random
-from cv2 import transform
-
-import albumentations as A
 import numpy as np
 
 import torch
 import torch.utils.data as Data
-import torch.distributed as dist
-from torch.utils.data.distributed import DistributedSampler
 from torch.utils.data import DataLoader
-from torchvision.transforms import transforms
 
-from simrec.datasets.transforms.randaug import RandAugment
-from simrec.utils.utils import label2yolobox
+from .utils import label2yolobox
 
 
 class RefCOCODataSet(Data.Dataset):
@@ -294,56 +287,3 @@ class RefCOCODataSet(Data.Dataset):
 
     def shuffle_list(self, list):
         random.shuffle(list)
-
-
-if __name__ == '__main__':
-
-    class Cfg():
-        def __init__(self):
-            super(Cfg, self).__init__()
-            self.ANN_PATH= {
-                'refcoco': './data/anns/refcoco.json',
-                'refcoco+': './data/anns/refcoco+.json',
-                'refcocog': './data/anns/refcocog.json',
-                'vg': './data/anns/vg.json',
-            }
-
-            self.IMAGE_PATH={
-                'refcoco': './data/images/train2014',
-                'refcoco+': './data/images/train2014',
-                'refcocog': './data/images/train2014',
-                'vg': './data/images/VG'
-            }
-
-            self.MASK_PATH={
-                'refcoco': './data/masks/refcoco',
-                'refcoco+': './data/masks/refcoco+',
-                'refcocog': './data/masks/refcocog',
-                'vg': './data/masks/vg'}
-            self.INPUT_SHAPE = (416, 416)
-            self.USE_GLOVE = True
-            self.DATASET = 'vg'
-            self.MAX_TOKEN = 15
-            self.MEAN = [0., 0., 0.]
-            self.STD = [1., 1., 1.]
-    cfg=Cfg()
-    dataset=RefCOCODataSet(cfg,'val')
-    data_loader = DataLoader(dataset,
-                             batch_size=10,
-                             shuffle=False,
-                             pin_memory=True)
-    for _,ref_iter,image_iter, mask_iter, box_iter,words,info_iter in data_loader:
-        print(ref_iter)
-        # print(image_iter.size())
-        # print(mask_iter.size())
-        # print(box_iter.size())
-        # print(ref_iter.size())
-        # # cv2.imwrite('./test.jpg', image_iter.numpy()[0].transpose((1, 2, 0))*255)
-        # # cv2.imwrite('./mask.jpg', mask_iter.numpy()[0].transpose((1, 2, 0))*255)
-        # print(info_iter.size())
-        # print(info_iter)
-
-
-
-
-
