+++ simrec/models/utils/nms.py
@@ -1,0 +1,135 @@
+# coding=utf-8
+# Copyright 2022 The SimREC Authors. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import numpy as np
+
+import torch
+
+
+def nms(bbox, thresh, score=None, limit=None):
+    """Suppress bounding boxes according to their IoUs and confidence scores.
+    Args:
+        bbox (array): Bounding boxes to be transformed. The shape is
+            :math:`(R, 4)`. :math:`R` is the number of bounding boxes.
+        thresh (float): Threshold of IoUs.
+        score (array): An array of confidences whose shape is :math:`(R,)`.
+        limit (int): The upper bound of the number of the output bounding
+            boxes. If it is not specified, this method selects as many
+            bounding boxes as possible.
+    Returns:
+        array:
+        An array with indices of bounding boxes that are selected. \
+        They are sorted by the scores of bounding boxes in descending \
+        order. \
+        The shape of this array is :math:`(K,)` and its dtype is\
+        :obj:`numpy.int32`. Note that :math:`K \\leq R`.
+
+    from: https://github.com/chainer/chainercv
+    """
+
+    if len(bbox) == 0:
+        return np.zeros((0,), dtype=np.int32)
+
+    if score is not None:
+        order = score.argsort()[::-1]
+        bbox = bbox[order]
+    bbox_area = np.prod(bbox[:, 2:] - bbox[:, :2], axis=1)
+
+    selec = np.zeros(bbox.shape[0], dtype=bool)
+    for i, b in enumerate(bbox):
+        tl = np.maximum(b[:2], bbox[selec, :2])
+        br = np.minimum(b[2:], bbox[selec, 2:])
+        area = np.prod(br - tl, axis=1) * (tl < br).all(axis=1)
+
+        iou = area / (bbox_area[i] + bbox_area[selec] - area)
+        if (iou >= thresh).any():
+            continue
+
+        selec[i] = True
+        if limit is not None and np.count_nonzero(selec) >= limit:
+            break
+
+    selec = np.where(selec)[0]
+    if score is not None:
+        selec = order[selec]
+    return selec.astype(np.int32)
+
+
+def postprocess(prediction, num_classes, conf_thre=0.7, nms_thre=0.45):
+    """
+    Postprocess for the output of YOLO model
+    perform box transformation, specify the class for each detection,
+    and perform class-wise non-maximum suppression.
+    Args:
+        prediction (torch tensor): The shape is :math:`(N, B, 4)`.
+            :math:`N` is the number of predictions,
+            :math:`B` the number of boxes. The last axis consists of
+            :math:`xc, yc, w, h` where `xc` and `yc` represent a center
+            of a bounding box.
+        num_classes (int):
+            number of dataset classes.
+        conf_thre (float):
+            confidence threshold ranging from 0 to 1,
+            which is defined in the config file.
+        nms_thre (float):
+            IoU threshold of non-max suppression ranging from 0 to 1.
+
+    Returns:
+        output (list of torch tensor):
+
+    """
+    box_corner = prediction.new(prediction.shape)
+    box_corner[:, :, 0] = prediction[:, :, 0] - prediction[:, :, 2] / 2
+    box_corner[:, :, 1] = prediction[:, :, 1] - prediction[:, :, 3] / 2
+    box_corner[:, :, 2] = prediction[:, :, 0] + prediction[:, :, 2] / 2
+    box_corner[:, :, 3] = prediction[:, :, 1] + prediction[:, :, 3] / 2
+    prediction[:, :, :4] = box_corner[:, :, :4]
+
+    output = [None for _ in range(len(prediction))]
+    for i, image_pred in enumerate(prediction):
+        # Filter out confidence scores below threshold
+        # class_pred = torch.max(image_pred[:, 5:5 + num_classes], 1)
+        # class_pred = class_pred[0]
+        conf_mask = (image_pred[:, 4]  >= conf_thre).squeeze()
+        image_pred = image_pred[conf_mask]
+
+        # If none are remaining => process next image
+        if not image_pred.size(0):
+            continue
+        # Get detections with higher confidence scores than the threshold
+        ind = (image_pred[:, 5:] * image_pred[:, 4][:, None] >= conf_thre).nonzero()
+        # Detections ordered as (x1, y1, x2, y2, obj_conf, class_conf, class_pred)
+        detections = torch.cat((
+                image_pred[ind[:, 0], :5],
+                image_pred[ind[:, 0], 5 + ind[:, 1]].unsqueeze(1),
+                ind[:, 1].float().unsqueeze(1)
+                ), 1)
+        # Iterate through all predicted classes
+        unique_labels = detections[:, -1].cpu().unique()
+        if prediction.is_cuda:
+            unique_labels = unique_labels.cuda()
+        for c in unique_labels:
+            # Get the detections with the particular class
+            detections_class = detections[detections[:, -1] == c]
+            nms_in = detections_class.cpu().numpy()
+            nms_out_index = nms(
+                nms_in[:, :4], nms_thre, score=nms_in[:, 4]*nms_in[:, 5])
+            detections_class = detections_class[nms_out_index]
+            if output[i] is None:
+                output[i] = detections_class
+            else:
+                output[i] = torch.cat((output[i], detections_class))
+
+    return output
