+++ simrec/utils/utils.py
@@ -189,34 +189,6 @@ def find_free_port():
     return port
 
 
-def lr_step_decay(lr_start=0.001, steps=[30, 40]):
-    def get_lr(epoch):
-        decay_rate = len(steps)
-        for i, e in enumerate(steps):
-            if epoch < e:
-                decay_rate = i
-                break
-        lr = lr_start / (10 ** (decay_rate))
-        return lr
-
-    return get_lr
-
-
-def lr_power_decay(lr_start=2.5e-4, lr_power=0.9, warm_up_lr=0., step_all=45 * 1414, warm_up_step=1000):
-    # step_per_epoch=3286
-    def warm_up(base_lr, lr, cur_step, end_step):
-        return base_lr + (lr - base_lr) * cur_step / end_step
-
-    def get_learningrate(epoch):
-
-        if epoch < warm_up_step:
-            lr = warm_up(warm_up_lr, lr_start, epoch, warm_up_step)
-        else:
-            lr = lr_start * ((1 - float(epoch - warm_up_step) / (step_all - warm_up_step)) ** lr_power)
-        return lr
-        # print("learning rate is", lr)
-
-    return get_learningrate
 
 def normed2original(image,mean=None,std=None,transpose=True):
     """
@@ -265,254 +237,7 @@ def draw_visualization(image,sent,pred_box,gt_box,draw_text=True,savepath=None):
                     0.8, colors[2], 2)
     return image
 
-def nms(bbox, thresh, score=None, limit=None):
-    """Suppress bounding boxes according to their IoUs and confidence scores.
-    Args:
-        bbox (array): Bounding boxes to be transformed. The shape is
-            :math:`(R, 4)`. :math:`R` is the number of bounding boxes.
-        thresh (float): Threshold of IoUs.
-        score (array): An array of confidences whose shape is :math:`(R,)`.
-        limit (int): The upper bound of the number of the output bounding
-            boxes. If it is not specified, this method selects as many
-            bounding boxes as possible.
-    Returns:
-        array:
-        An array with indices of bounding boxes that are selected. \
-        They are sorted by the scores of bounding boxes in descending \
-        order. \
-        The shape of this array is :math:`(K,)` and its dtype is\
-        :obj:`numpy.int32`. Note that :math:`K \\leq R`.
 
-    from: https://github.com/chainer/chainercv
-    """
-
-    if len(bbox) == 0:
-        return np.zeros((0,), dtype=np.int32)
-
-    if score is not None:
-        order = score.argsort()[::-1]
-        bbox = bbox[order]
-    bbox_area = np.prod(bbox[:, 2:] - bbox[:, :2], axis=1)
-
-    selec = np.zeros(bbox.shape[0], dtype=bool)
-    for i, b in enumerate(bbox):
-        tl = np.maximum(b[:2], bbox[selec, :2])
-        br = np.minimum(b[2:], bbox[selec, 2:])
-        area = np.prod(br - tl, axis=1) * (tl < br).all(axis=1)
-
-        iou = area / (bbox_area[i] + bbox_area[selec] - area)
-        if (iou >= thresh).any():
-            continue
-
-        selec[i] = True
-        if limit is not None and np.count_nonzero(selec) >= limit:
-            break
-
-    selec = np.where(selec)[0]
-    if score is not None:
-        selec = order[selec]
-    return selec.astype(np.int32)
-
-
-def postprocess(prediction, num_classes, conf_thre=0.7, nms_thre=0.45):
-    """
-    Postprocess for the output of YOLO model
-    perform box transformation, specify the class for each detection,
-    and perform class-wise non-maximum suppression.
-    Args:
-        prediction (torch tensor): The shape is :math:`(N, B, 4)`.
-            :math:`N` is the number of predictions,
-            :math:`B` the number of boxes. The last axis consists of
-            :math:`xc, yc, w, h` where `xc` and `yc` represent a center
-            of a bounding box.
-        num_classes (int):
-            number of dataset classes.
-        conf_thre (float):
-            confidence threshold ranging from 0 to 1,
-            which is defined in the config file.
-        nms_thre (float):
-            IoU threshold of non-max suppression ranging from 0 to 1.
-
-    Returns:
-        output (list of torch tensor):
-
-    """
-    box_corner = prediction.new(prediction.shape)
-    box_corner[:, :, 0] = prediction[:, :, 0] - prediction[:, :, 2] / 2
-    box_corner[:, :, 1] = prediction[:, :, 1] - prediction[:, :, 3] / 2
-    box_corner[:, :, 2] = prediction[:, :, 0] + prediction[:, :, 2] / 2
-    box_corner[:, :, 3] = prediction[:, :, 1] + prediction[:, :, 3] / 2
-    prediction[:, :, :4] = box_corner[:, :, :4]
-
-    output = [None for _ in range(len(prediction))]
-    for i, image_pred in enumerate(prediction):
-        # Filter out confidence scores below threshold
-        # class_pred = torch.max(image_pred[:, 5:5 + num_classes], 1)
-        # class_pred = class_pred[0]
-        conf_mask = (image_pred[:, 4]  >= conf_thre).squeeze()
-        image_pred = image_pred[conf_mask]
-
-        # If none are remaining => process next image
-        if not image_pred.size(0):
-            continue
-        # Get detections with higher confidence scores than the threshold
-        ind = (image_pred[:, 5:] * image_pred[:, 4][:, None] >= conf_thre).nonzero()
-        # Detections ordered as (x1, y1, x2, y2, obj_conf, class_conf, class_pred)
-        detections = torch.cat((
-                image_pred[ind[:, 0], :5],
-                image_pred[ind[:, 0], 5 + ind[:, 1]].unsqueeze(1),
-                ind[:, 1].float().unsqueeze(1)
-                ), 1)
-        # Iterate through all predicted classes
-        unique_labels = detections[:, -1].cpu().unique()
-        if prediction.is_cuda:
-            unique_labels = unique_labels.cuda()
-        for c in unique_labels:
-            # Get the detections with the particular class
-            detections_class = detections[detections[:, -1] == c]
-            nms_in = detections_class.cpu().numpy()
-            nms_out_index = nms(
-                nms_in[:, :4], nms_thre, score=nms_in[:, 4]*nms_in[:, 5])
-            detections_class = detections_class[nms_out_index]
-            if output[i] is None:
-                output[i] = detections_class
-            else:
-                output[i] = torch.cat((output[i], detections_class))
-
-    return output
-
-
-def bboxes_iou(bboxes_a, bboxes_b, xyxy=True):
-    """Calculate the Intersection of Unions (IoUs) between bounding boxes.
-    IoU is calculated as a ratio of area of the intersection
-    and area of the union.
-
-    Args:
-        bbox_a (array): An array whose shape is :math:`(N, 4)`.
-            :math:`N` is the number of bounding boxes.
-            The dtype should be :obj:`numpy.float32`.
-        bbox_b (array): An array similar to :obj:`bbox_a`,
-            whose shape is :math:`(K, 4)`.
-            The dtype should be :obj:`numpy.float32`.
-    Returns:
-        array:
-        An array whose shape is :math:`(N, K)`. \
-        An element at index :math:`(n, k)` contains IoUs between \
-        :math:`n` th bounding box in :obj:`bbox_a` and :math:`k` th bounding \
-        box in :obj:`bbox_b`.
-
-    from: https://github.com/chainer/chainercv
-    """
-    if bboxes_a.shape[1] != 4 or bboxes_b.shape[1] != 4:
-        raise IndexError
-
-    # top left
-    if xyxy:
-        tl = torch.max(bboxes_a[:, None, :2], bboxes_b[:, :2])
-        # bottom right
-        br = torch.min(bboxes_a[:, None, 2:], bboxes_b[:, 2:])
-        area_a = torch.prod(bboxes_a[:, 2:] - bboxes_a[:, :2], 1)
-        area_b = torch.prod(bboxes_b[:, 2:] - bboxes_b[:, :2], 1)
-    else:
-        tl = torch.max((bboxes_a[:, None, :2] - bboxes_a[:, None, 2:] / 2),
-                        (bboxes_b[:, :2] - bboxes_b[:, 2:] / 2))
-        # bottom right
-        br = torch.min((bboxes_a[:, None, :2] + bboxes_a[:, None, 2:] / 2),
-                        (bboxes_b[:, :2] + bboxes_b[:, 2:] / 2))
-
-        area_a = torch.prod(bboxes_a[:, 2:], 1)
-        area_b = torch.prod(bboxes_b[:, 2:], 1)
-    en = (tl < br).type(tl.type()).prod(dim=2)
-    area_i = torch.prod(br - tl, 2) * en  # * ((tl < br).all())
-    return area_i / (area_a[:, None] + area_b - area_i)
-
-
-def batch_box_iou(box1, box2,threshold=0.5):
-    """
-    :param box1:  N,4
-    :param box2:  N,4
-    :return: N
-    """
-    in_h = torch.min(box1[:,2], box2[:,2]) - torch.max(box1[:,0], box2[:,0])
-    in_w = torch.min(box1[:,3], box2[:,3]) - torch.max(box1[:,1], box2[:,1])
-    in_h=in_h.clamp(min=0.)
-    in_w=in_w.clamp(min=0.)
-    inter =in_h * in_w
-    union = (box1[:,2] - box1[:,0]) * (box1[:,3] - box1[:,1]) + \
-            (box2[:,2] - box2[:,0]) * (box2[:,3] - box2[:,1]) - inter
-    iou = inter / union
-    return iou>threshold
-
-
-def mask_iou(mask1, mask2):
-    """
-    :param mask1:  l
-    :param mask2:  l
-    :return: iou
-    """
-    mask1 =mask1.reshape([-1])
-    mask2=mask2.reshape([-1])
-    t = np.array(mask1 > 0.5)
-    p = mask2 > 0.
-    intersection = np.logical_and(t, p)
-    union = np.logical_or(t, p)
-    iou = (np.sum(intersection > 0) + 1e-10) / (np.sum(union > 0) + 1e-10)
-
-    ap = dict()
-    thresholds = np.arange(0.5, 1, 0.05)
-    s = []
-    for thresh in thresholds:
-        ap[thresh] = float(iou > thresh)
-    return iou,ap
-def mask_processing(mask,info_img):
-    # print(info_img)
-    h, w, nh, nw, dx, dy,_=info_img
-    # print(info_img)
-    # print(mask)
-    mask=mask[dy:dy + nh, dx:dx + nw,None]
-    mask=cv2.resize(mask,(int(w),int(h)))
-    return mask
-
-
-
-
-def label2yolobox(labels, info_img, maxsize, lrflip=False):
-    """
-    Transform coco labels to yolo box labels
-    Args:
-        labels (numpy.ndarray): label data whose shape is :math:`(N, 5)`.
-            Each label consists of [class, x, y, w, h] where \
-                class (float): class index.
-                x, y, w, h (float) : coordinates of \
-                    left-top points, width, and height of a bounding box.
-                    Values range from 0 to width or height of the image.
-        info_img : tuple of h, w, nh, nw, dx, dy.
-            h, w (int): original shape of the image
-            nh, nw (int): shape of the resized image without padding
-            dx, dy (int): pad size
-        maxsize (int): target image size after pre-processing
-        lrflip (bool): horizontal flip flag
-
-    Returns:
-        labels:label data whose size is :math:`(N, 5)`.
-            Each label consists of [class, xc, yc, w, h] where
-                class (float): class index.
-                xc, yc (float) : center of bbox whose values range from 0 to 1.
-                w, h (float) : size of bbox whose values range from 0 to 1.
-    """
-    h, w, nh, nw, dx, dy,_ = info_img
-    x1 = labels[:, 0] / w
-    y1 = labels[:, 1] / h
-    x2 = (labels[:, 0] + labels[:, 2]) / w
-    y2 = (labels[:, 1] + labels[:, 3]) / h
-    labels[:, 0] = (((x1 + x2) / 2) * nw + dx) / maxsize
-    labels[:, 1] = (((y1 + y2) / 2) * nh + dy) / maxsize
-    labels[:, 2] *= nw / w / maxsize
-    labels[:, 3] *= nh / h / maxsize
-    labels[:,:4]=np.clip(labels[:,:4],0.,0.99)
-    if lrflip:
-        labels[:, 0] = 1 - labels[:, 0]
-    return labels
 
 
 def yolobox2label(box, info_img):
